requirements.txt
# langchain==0.0.353   # required
# pip install -U langchain-community
# pip install -U langchain-openai
# openai==1.3.5   # required
# pydantic==2.4.2   # required
# python-dotenv==1.0.0   # required
# requests==2.31.0   # required

# fastapi==0.108.0 .  # required
# # pip install "uvicorn[standard]" .  # required
# Jinja2==3.1.2   # required

aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.3
distro==1.9.0
fastapi==0.109.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.32
langchain-openai==0.0.5
langchain-text-splitters==0.0.1
langsmith==0.1.27
MarkupSafe==2.1.4
marshmallow==3.20.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.3
openai==1.10.0
orjson==3.9.15
packaging==23.2
pydantic==2.6.0
pydantic_core==2.16.1
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.36.3
tenacity==8.2.3
tiktoken==0.5.2
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
